Started a config file which tells docker how to config the network and where the dockerfile is
