+++ config/mongodb.json
@@ -1,0 +1,12 @@
+{
+  "dockerfile": "MongoDB",
+  "containerName": "my_mongodb_container",
+  "dockerFileName": "mongoDBImage",
+  "exposedPorts": {"27017/tcp": {}},
+  "HostConfig": { "PortBindings": {"27017/tcp": [{"HostPort": "27017"}]}},
+  "network": "bridge",
+  "arguments": {
+    "ENV_VAR1": "value1",
+    "ENV_VAR2": "value2"
+  }
+}
